refactor: 테스트 디렉토리의 PointService, DefaultPointService의 내용을 common, user모듈 클래스로 이동
