+++ user/src/main/java/com/food/user/business/DefaultPointService.java
@@ -1,96 +1,71 @@
 package com.food.user.business;
 
+import com.food.common.payment.business.external.PaymentAmountService;
 import com.food.common.user.business.external.PointService;
-import com.food.common.user.business.external.model.PointCollectRequest;
-import com.food.common.user.business.external.model.PointUseRequest;
+import com.food.common.user.business.external.model.RequestUser;
 import com.food.common.user.business.internal.PointCommonService;
+import com.food.common.user.business.internal.UserCommonService;
 import com.food.common.user.business.internal.dto.PointDto;
-import com.food.common.user.business.internal.dto.PointSaveDto;
-import com.food.common.user.enumeration.PointType;
+import com.food.common.user.business.internal.dto.UserDto;
 import com.food.common.utils.Amount;
-import lombok.RequiredArgsConstructor;
+import com.food.common.utils.UsedPoints;
+import com.food.user.error.InsufficientPointBalanceException;
+import com.food.user.error.NotFoundPointOwnerException;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-import javax.validation.constraints.NotNull;
-import java.util.List;
-import java.util.Optional;
-
-@RequiredArgsConstructor
-@Transactional
 @Service
 public class DefaultPointService implements PointService {
+    private final UserCommonService userCommonService;
     private final PointCommonService pointCommonService;
+    private final PaymentAmountService paymentAmountService;
 
-    @Override
-    public Long use(PointUseRequest request) {
-        Amount currentAmount = currentAmount(request.getOwnerId());
-        if (request.hasGreaterAmountThan(currentAmount)) {
-            throw new IllegalArgumentException("잔여포인트가 부족합니다. 잔여포인트=" + currentAmount);
-        }
-
-        return pointCommonService.save(request.toPointSaveDto());
+    public DefaultPointService(UserCommonService userCommonService, PointCommonService pointCommonService, PaymentAmountService paymentAmountService) {
+        this.userCommonService = userCommonService;
+        this.pointCommonService = pointCommonService;
+        this.paymentAmountService = paymentAmountService;
     }
 
     @Override
-    public void collect(PointCollectRequest request) {
-        int collectAmount = calculateCollectAmount(0.05F, request.getPaymentAmount());
+    public void collect(Long paymentId, RequestUser requestUser) {
+        UserDto owner = userCommonService.findById(requestUser.getUserId())
+                .orElseThrow(() -> new NotFoundPointOwnerException(requestUser.getUserId()));
 
-        pointCommonService.save(request.toPointSaveDto(collectAmount));
+        Amount paymentAmount = paymentAmountService.sumOfPaymentAmountWithoutPoints(paymentId);
+        Amount collectedAmount = paymentAmount.multiply(0.05f);
+
+        PointDto collectPoint = basePoint(owner.getId())
+                .collect(collectedAmount, paymentId);
+
+        pointCommonService.save(collectPoint);
+    }
+
+    @Override
+    public Long use(UsedPoints amount, RequestUser requestUser) {
+        if (!userCommonService.existsById(requestUser.getUserId())) {
+            throw new NotFoundPointOwnerException(requestUser.getUserId());
+        }
+
+        PointDto basePoint = basePoint(requestUser.getUserId());
+        if (!basePoint.getCurrentAmount().isGreaterThanOrEqualTo(amount)) {
+            throw new InsufficientPointBalanceException();
+        }
+
+        PointDto usePoint = basePoint.use(amount);
+        return pointCommonService.save(usePoint).getId();
     }
 
     @Override
     public void recollectUsedPoint(Long pointId) {
-        Optional<PointDto> optionalPoint = pointCommonService.findByPointId(pointId);
-        if (optionalPoint.isEmpty()) return;
 
-        PointDto point = optionalPoint.get();
-        PointSaveDto request = PointSaveDto.builder()
-                .usedId(point.getUserId())
-                .amount(point.getChangedAmount().getValue())
-                .type(PointType.RECOLLECT)
-                .paymentId(point.getPaymentId())
-                .build();
-
-        pointCommonService.save(request);
     }
 
     @Override
     public void retrieveCollectedPoint(Long paymentId) {
-        List<PointDto> pointHistory = pointCommonService.findAllByPaymentId(paymentId);
-        PointDto point = getValidatedRetrievedPoint(pointHistory);
 
-        PointSaveDto request = PointSaveDto.builder()
-                .usedId(point.getUserId())
-                .amount(point.getChangedAmount().getValue())
-                .type(PointType.RETRIEVE)
-                .paymentId(paymentId)
-                .build();
-
-        pointCommonService.save(request);
     }
 
-    private PointDto getValidatedRetrievedPoint(List<PointDto> pointHistory) {
-        boolean existsRetrieveHistory = pointHistory.stream()
-                .anyMatch(point -> point.getType() == PointType.RETRIEVE);
-
-        if (existsRetrieveHistory) {
-            throw new IllegalArgumentException("이미 회수내역이 존재합니다.");
-        }
-
-        return pointHistory.stream()
-                .filter(point -> point.getType() == PointType.COLLECT)
-                .findAny()
-                .orElseThrow(() -> new IllegalArgumentException("적립내역이 존재하지 않습니다."));
-    }
-
-    private int calculateCollectAmount(float accumulationRate, int paymentAmount) {
-        return (int) Math.ceil(paymentAmount * accumulationRate);
-    }
-
-    private Amount currentAmount(@NotNull Long userId) {
-        return pointCommonService.findLatestPointByUserId(userId)
-                .map(PointDto::getCurrentAmount)
-                .orElse(Amount.zero());
+    private PointDto basePoint(Long ownerId) {
+        return pointCommonService.findLatestPointByUserId(ownerId)
+                .orElse(PointDto.createBasePoint(ownerId));
     }
 }
